Move watchdog to util namespace
